# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Celljevity Longevity OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List patients
      parameters:
        - name: stage
          in: query
          schema:
            $ref: "#/components/schemas/JourneyStage"
        - name: isLead
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Patient list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientListResponse"
...
  /patients/{patientId}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient detail
      parameters:
        - $ref: "#/components/parameters/PatientId"
      responses:
        "200":
          description: Patient detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientDetail"
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update patient
      parameters:
        - $ref: "#/components/parameters/PatientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatientRequest"
      responses:
        "200":
          description: Patient updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List quotes
      parameters:
        - name: patientId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/QuoteStatus"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Quote list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteListResponse"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Create quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteRequest"
      responses:
        "201":
          description: Quote created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{quoteId}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get quote with line items
      parameters:
        - $ref: "#/components/parameters/QuoteId"
      responses:
        "200":
          description: Quote detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteDetail"
    patch:
      operationId: updateQuote
      tags: [quotes]
      summary: Update quote
      parameters:
        - $ref: "#/components/parameters/QuoteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuoteRequest"
      responses:
        "200":
          description: Quote updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{quoteId}/line-items:
    post:
      operationId: addLineItem
      tags: [quotes]
      summary: Add line item to quote
      parameters:
        - $ref: "#/components/parameters/QuoteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLineItemRequest"
      responses:
        "201":
          description: Line item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LineItem"
  /quotes/{quoteId}/line-items/{lineItemId}:
    delete:
      operationId: deleteLineItem
      tags: [quotes]
      summary: Delete line item
      parameters:
        - $ref: "#/components/parameters/QuoteId"
        - name: lineItemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Line item deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  parameters:
    PatientId:
      name: patientId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    QuoteId:
      name: quoteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    JourneyStage:
      type: string
      enum: [ACQUISITION, INTAKE, DIAGNOSTICS, PLANNING, TREATMENT, FOLLOW_UP]
    ServiceCategory:
      type: string
      enum: [EXOSOMES, PROMETHEUS, NK_CELLS, DIAGNOSTICS, OTHER]
    QuoteStatus:
      type: string
      enum: [DRAFT, PENDING_SIGNATURE, ACCEPTED, PAID, CANCELLED]
...
    Patient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        celljevityId:
          type: string
        dateOfBirth:
          type: string
          format: date
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        journeyStage:
          $ref: "#/components/schemas/JourneyStage"
        medicalHistorySummary:
          type: string
          nullable: true
        assignedCoordinatorId:
          type: string
          format: uuid
          nullable: true
        assignedProviderId:
          type: string
          format: uuid
          nullable: true
        isLead:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, celljevityId, journeyStage, isLead, createdAt, updatedAt]
    PatientDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        celljevityId:
          type: string
        journeyStage:
          $ref: "#/components/schemas/JourneyStage"
        isLead:
          type: boolean
        dateOfBirth:
          type: string
          format: date
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        medicalHistorySummary:
          type: string
          nullable: true
        assignedCoordinatorId:
          type: string
          format: uuid
          nullable: true
        assignedProviderId:
          type: string
          format: uuid
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, celljevityId, journeyStage, isLead, firstName, lastName, email, createdAt, updatedAt]
    PatientListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PatientDetail"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [data, total, limit, offset]
    UpdatePatientRequest:
      type: object
      properties:
        journeyStage:
          $ref: "#/components/schemas/JourneyStage"
        isLead:
          type: boolean
        assignedCoordinatorId:
          type: string
          format: uuid
          nullable: true
        assignedProviderId:
          type: string
          format: uuid
          nullable: true
        dateOfBirth:
          type: string
          format: date
        phone:
          type: string
        address:
          type: string
        medicalHistorySummary:
          type: string
...
    Quote:
      type: object
      properties:
        id:
          type: string
          format: uuid
        invoiceNumber:
          type: string
        patientId:
          type: string
          format: uuid
        createdBy:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/QuoteStatus"
        issueDate:
          type: string
          format: date
        currency:
          type: string
        exchangeRateUsed:
          type: string
        totalAmount:
          type: string
        notesAndTerms:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, invoiceNumber, patientId, createdBy, status, issueDate, currency, exchangeRateUsed, totalAmount, createdAt, updatedAt]
    LineItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        quoteId:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        customDescription:
          type: string
          nullable: true
        quantity:
          type: integer
        unitPrice:
          type: string
        lineTotal:
          type: string
      required: [id, quoteId, serviceId, quantity, unitPrice, lineTotal]
    QuoteDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        invoiceNumber:
          type: string
        patientId:
          type: string
          format: uuid
        createdBy:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/QuoteStatus"
        issueDate:
          type: string
          format: date
        currency:
          type: string
        exchangeRateUsed:
          type: string
        totalAmount:
          type: string
        notesAndTerms:
          type: string
          nullable: true
        lineItems:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              serviceId:
                type: string
                format: uuid
              customDescription:
                type: string
                nullable: true
              quantity:
                type: integer
              unitPrice:
                type: string
              lineTotal:
                type: string
              serviceName:
                type: string
                nullable: true
              serviceCategory:
                $ref: "#/components/schemas/ServiceCategory"
            required: [id, serviceId, quantity, unitPrice, lineTotal]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, invoiceNumber, patientId, createdBy, status, issueDate, currency, totalAmount, lineItems, createdAt, updatedAt]
    QuoteListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Quote"
        limit:
          type: integer
        offset:
          type: integer
      required: [data, limit, offset]
    CreateQuoteRequest:
      type: object
      properties:
        patientId:
          type: string
          format: uuid
        currency:
          type: string
        exchangeRateUsed:
          type: number
        notesAndTerms:
          type: string
        issueDate:
          type: string
          format: date
      required: [patientId]
    UpdateQuoteRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/QuoteStatus"
        currency:
          type: string
        exchangeRateUsed:
          type: number
        notesAndTerms:
          type: string
        totalAmount:
          type: number
    CreateLineItemRequest:
      type: object
      properties:
        serviceId:
          type: string
          format: uuid
        customDescription:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
      required: [serviceId, unitPrice]
...